+++ tests/protocols/gmpv208/entities/secinfo/test_get_nvt.py
@@ -16,16 +16,29 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from gvm.errors import RequiredArgument
 
-class GmpGetNvtFamiliesTestCase:
-    def test_get_nvt_families(self):
-        self.gmp.get_nvt_families()
 
-        self.connection.send.has_been_called_with('<get_nvt_families/>')
-
-    def test_get_nvt_families_with_sort_order(self):
-        self.gmp.get_nvt_families(sort_order='foo')
+class GmpGetNvtTestMixin:
+    def test_get_nvt(self):
+        self.gmp.get_nvt(nvt_id='i1')
 
         self.connection.send.has_been_called_with(
-            '<get_nvt_families sort_order="foo"/>'
+            '<get_info info_id="i1" type="NVT" details="1"/>'
         )
+
+        self.gmp.get_nvt('i1')
+
+        self.connection.send.has_been_called_with(
+            '<get_info info_id="i1" type="NVT" details="1"/>'
+        )
+
+    def test_get_nvt_missing_nvt_id(self):
+        with self.assertRaises(RequiredArgument):
+            self.gmp.get_nvt(nvt_id='')
+
+        with self.assertRaises(RequiredArgument):
+            self.gmp.get_nvt('')
+
+        with self.assertRaises(RequiredArgument):
+            self.gmp.get_nvt(nvt_id=None)