+++ CHANGELOG.md
@@ -10,11 +10,16 @@ and this project adheres to [Calendar Versioning](https://calver.org)html).
 **Dropped support for GMP v7, v8 and v9!**
 
 ### Added
+* Introduced new explicit API calls for SecInfo: `get_nvt()`, `get_nvt_list()`, `get_cpe()`, `get_cpe_list()`, `get_cve()`, `get_cve_list()`, `get_cert_bund_advisory()`, `get_cert_bund_advisory_list()`, `get_dnf_cert_advisory()`, `get_dnf_cert_advisory_list()`, `get_oval_definition()`, `get_oval_definition_list()`. [#456](https://github.com/greenbone/python-gvm/pull/456)
+
 ### Changed
+* Changed the API calls `get_nvt()` and `get_nvts()` to `get_scan_config_nvt()` and `get_scan_config_nvts()`.  [#456](https://github.com/greenbone/python-gvm/pull/456)
+* Detached the `InfoType` from the GMP types class.  [#456](https://github.com/greenbone/python-gvm/pull/456)
+* Detached the SecInfo (CPE, CVE, NVT, CERT-BUND, DNF-CERT, OVAL Definitions) calls from GMP class into new `SecInfoMixin`. [#456](https://github.com/greenbone/python-gvm/pull/456)
 * Detached the PortList and PortRange API calls from the GMP class into a new `PortListMixin`. [#446](https://github.com/greenbone/python-gvm/pull/446)
 * Detached the Target API calls from the GMP class into a new `TargetMixin`. [#446](https://github.com/greenbone/python-gvm/pull/446)
-* Detached the AliveTest from the GMP types class. [#446](https://github.com/greenbone/python-gvm/pull/446)
-* Detached the PortListType from the GMP types class. [#446](https://github.com/greenbone/python-gvm/pull/446)
+* Detached the `AliveTest` from the GMP types class. [#446](https://github.com/greenbone/python-gvm/pull/446)
+* Detached the `PortListType` from the GMP types class. [#446](https://github.com/greenbone/python-gvm/pull/446)
 * Detached the ReportFormatType from the GMP types class. [#445](https://github.com/greenbone/python-gvm/pull/445)
 * Detached the Report API calls from the GMP class into a new `ReportMixin`. [#445](https://github.com/greenbone/python-gvm/pull/445)
 * Detached the Task API calls from the GMP class into a new `TaskMixin`. [#443](https://github.com/greenbone/python-gvm/pull/443)